DRTII-1254 Tidying
