+++ server/src/test/scala/services/exports/PassengerExportsSpec.scala
@@ -6,7 +6,7 @@ import controllers.ArrivalGenerator
 import drt.shared.CrunchApi.PassengersMinute
 import passengersplits.parsing.VoyageManifestParser.VoyageManifests
 import services.crunch.CrunchTestLike
-import uk.gov.homeoffice.drt.arrivals.{ApiFlightWithSplits, Arrival, ArrivalStatus, CarrierCode, FlightsWithSplits, Passengers, VoyageNumber}
+import uk.gov.homeoffice.drt.arrivals._
 import uk.gov.homeoffice.drt.ports.Queues.{EGate, EeaDesk, NonEeaDesk, Queue}
 import uk.gov.homeoffice.drt.ports.Terminals.{T1, Terminal}
 import uk.gov.homeoffice.drt.ports._
@@ -16,52 +16,36 @@ import scala.collection.immutable.Seq
 import scala.concurrent.duration.DurationInt
 import scala.concurrent.{Await, Future}
 
-class FlightExportsSpec extends CrunchTestLike {
+class PassengerExportsSpec extends CrunchTestLike {
   val paxSourceOrder: List[FeedSource] = List(LiveFeedSource)
 
   def passengers(maybeActualPax: Option[Int], maybeTransPax: Option[Int]): Map[FeedSource, Passengers] =
     Map[FeedSource, Passengers](LiveFeedSource -> Passengers(maybeActualPax, maybeTransPax))
 
-  "toCsv should give a row for each flight relevant to the date range, including the region, port and terminal" >> {
-    val start = LocalDate(2020, 6, 1)
-    val end = LocalDate(2020, 6, 2)
-    val port = PortCode("MAN")
-    val terminal = Terminal("T1")
+  "paxForMinute should" >> {
+    "give 20 for minutes 1 to 4, 15 for minute 5, and 0 for minutes 0 & 6, given 95 total pax" >> {
+      (0 to 6).map(PassengerExports.paxForMinute(95, _)) === Seq(0, 20, 20, 20, 20, 15, 0)
+    }
+  }
 
-    val utcFlightsProvider: (UtcDate, UtcDate, Terminal) => Source[(UtcDate, FlightsWithSplits), NotUsed] = (_, _, _) =>
-      Source(List(
-        (UtcDate(2020, 6, 1), FlightsWithSplits(Seq(
-          ApiFlightWithSplits(ArrivalGenerator.arrival(
-            iata = "BA0001", schDt = "2020-06-01T20:00", pcpDt = "2020-06-02T01:30", passengerSources = passengers(Option(95), None)), Set()),
-        ))),
-        (UtcDate(2020, 6, 2), FlightsWithSplits(Seq(
-          ApiFlightWithSplits(ArrivalGenerator.arrival(
-            iata = "BA0002", schDt = "2020-06-02T00:05", pcpDt = "2020-06-02T00:30", passengerSources = passengers(Option(95), None)), Set()),
-        ))),
-        (UtcDate(2020, 6, 3), FlightsWithSplits(Seq(
-          ApiFlightWithSplits(ArrivalGenerator.arrival(
-            iata = "BA0003", schDt = "2020-06-03T00:05", pcpDt = "2020-06-02T22:55", passengerSources = passengers(Option(95), None)), Set()),
-        ))),
-        (UtcDate(2020, 6, 4), FlightsWithSplits(Seq(
-          ApiFlightWithSplits(ArrivalGenerator.arrival(
-            iata = "BA0004", schDt = "2020-06-03T02:30", pcpDt = "2020-06-03T01:55", passengerSources = passengers(Option(95), None)), Set()),
-        ))),
-      ))
+  "queueTotals should" >> {
+    "give sums of passengers for each queue in a collection of CrunchMinutes" >> {
+      def paxMinute(minute: Int, queue: Queue, pax: Int): PassengersMinute =
+        PassengersMinute(T1, queue, minute * oneMinuteMillis, Iterable.fill(pax)(0), None)
 
-    "Given a flights provider, and dateAndFlightsToCsvRows as an aggregator" >> {
-      val getFlights = FlightExports.flightsProvider(utcFlightsProvider, paxSourceOrder)
-      val toRows = FlightExports.dateAndFlightsToCsvRows(port, terminal, paxSourceOrder, (_, _) => Future.successful(VoyageManifests.empty))
-      val csvStream = GeneralExport.toCsv(start, end, terminal, getFlights, toRows)
-
-      val result = Await.result(csvStream.runWith(Sink.seq), 1.second)
-      val expected = List(
-        """North,MAN,T1,BA0002,BA0002,JFK,/,Scheduled,2020-06-02 01:05,,,,,,,2020-06-02 01:30,95,95,,,,,,,,,,,,,,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"",""
-          |North,MAN,T1,BA0001,BA0001,JFK,/,Scheduled,2020-06-01 21:00,,,,,,,2020-06-02 02:30,95,95,,,,,,,,,,,,,,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"",""
-          |North,MAN,T1,BA0003,BA0003,JFK,/,Scheduled,2020-06-03 01:05,,,,,,,2020-06-02 23:55,95,95,,,,,,,,,,,,,,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"",""
-          |""".stripMargin
+      val cms = Seq(
+        paxMinute(1, Queues.EeaDesk, 1),
+        paxMinute(2, Queues.EeaDesk, 5),
+        paxMinute(1, Queues.EGate, 2),
+        paxMinute(2, Queues.EGate, 6),
+        paxMinute(1, Queues.NonEeaDesk, 3),
+        paxMinute(2, Queues.NonEeaDesk, 7),
       )
-
-      result === expected
+      PassengerExports.queueTotals(cms) === Map(
+        Queues.EeaDesk -> 6,
+        Queues.EGate -> 8,
+        Queues.NonEeaDesk -> 10,
+      )
     }
   }
